# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Freitas Outlet API
  description: API do e-commerce Freitas Outlet — streetwear, luxo & performance
  version: 0.1.0
  contact:
    name: Freitas Outlet
servers:
  - url: http://localhost:3000
    description: Desenvolvimento
  - url: https://freitasoutlet.com.br
    description: Produção
paths:
  /api/auth/register:
    post:
      summary: Registrar novo cliente
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string, minLength: 6 }
      responses:
        '200': { description: Registro realizado }
        '409': { description: E-mail já cadastrado }
  /api/auth/login:
    post:
      summary: Login do cliente
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Login realizado }
        '401': { description: Credenciais inválidas }
...
  /api/auth/forgot-password:
    post:
      summary: Solicitar recuperação de senha
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200': { description: E-mail enviado se conta existir }
  /api/auth/reset-password:
    post:
      summary: Redefinir senha
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string }
      responses:
        '200': { description: Senha redefinida }
...
  /api/admin/login:
    post:
      summary: Login administrativo
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
      responses:
        '200': { description: Login realizado }
...
  /api/admin/produtos:
    get:
      summary: Listar produtos
      tags: [Admin]
      responses:
        '200': { description: Lista de produtos }
    post:
      summary: Criar produto
      tags: [Admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
                price: { type: number }
                description: { type: string }
      responses:
        '200': { description: Produto criado }
...